requirements.txt
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow
pytesseract
httpx
matplotlib